Change package struecture - Split by layer -> domain
