+++ app/src/main/java/com/codesoom/assignment/App.java
@@ -1,9 +1,10 @@
 package com.codesoom.assignment;
 
-import com.codesoom.assignment.service.UserService;
+import com.codesoom.assignment.application.user.UserService;
 import com.codesoom.assignment.web.*;
-import com.codesoom.assignment.service.TaskService;
-import com.sun.net.httpserver.HttpHandler;
+import com.codesoom.assignment.application.task.TaskService;
+import com.codesoom.assignment.web.task.TaskHttpRequestContext;
+import com.codesoom.assignment.web.user.UserHttpRequestContext;
 
 import java.io.IOException;
