+++ app/src/main/java/com/codesoom/assignment/domain/Task.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.models;
+package com.codesoom.assignment.domain;
 
 public class Task {
     private long id;
