+++ app/src/main/java/com/codesoom/assignment/web/RequestControllable.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.web;
+
+import java.io.IOException;
+
+public interface RequestControllable {
+    HttpResponse process(HttpRequest httpRequest) throws IOException;
+}
